# docs/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Unified Authentication Business Shell API
  version: 0.1.0
servers:
  - url: /
security: []
paths:
...
  /v1/auth/callback:
    get:
      operationId: completeLogin
      summary: Handle the authorization code callback
      parameters:
        - in: query
          name: code
          required: true
          schema: { type: string }
        - in: query
          name: state
          required: true
          schema: { type: string }
      responses:
        '302': { description: Redirect to the configured web origin }
        '401':
          description: Authentication failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /v1/auth/refresh:
    post:
      operationId: refreshSession
      summary: Refresh the server-side Confidential Client session
      security:
        - sessionCookie: []
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema: { type: string }
      responses:
        '200': { description: Session refreshed }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: auth_template_session
  responses:
    Unauthorized:
      description: Missing or expired session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Authenticated caller lacks the required permission
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                required: [field, reason]
                properties:
                  field: { type: string }
                  reason: { type: string }
            requestId: { type: string }